refactor: move `ConversationEventReceiver` and `UserEventReceiver` to their own files
