+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/sync/UserEventReceiver.kt
@@ -4,7 +4,8 @@ import com.wire.kalium.logic.data.connection.ConnectionRepository
 import com.wire.kalium.logic.data.event.Event
 import com.wire.kalium.logic.functional.onFailure
 import com.wire.kalium.logic.kaliumLogger
-import com.wire.kalium.logic.sync.event.UserEventReceiver
+
+interface UserEventReceiver : EventReceiver<Event.User>
 
 class UserEventReceiverImpl(
     private val connectionRepository: ConnectionRepository,
