+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/sync/event/EventProcessor.kt
@@ -3,7 +3,8 @@ package com.wire.kalium.logic.sync.event
 import com.wire.kalium.logic.data.event.Event
 import com.wire.kalium.logic.data.event.EventRepository
 import com.wire.kalium.logic.kaliumLogger
-import com.wire.kalium.logic.sync.EventReceiver
+import com.wire.kalium.logic.sync.ConversationEventReceiver
+import com.wire.kalium.logic.sync.UserEventReceiver
 
 /**
  * Handles incoming events from remote.
@@ -17,9 +18,6 @@ internal interface EventProcessor {
      */
     suspend fun processEvent(event: Event)
 }
-
-interface ConversationEventReceiver : EventReceiver<Event.Conversation>
-interface UserEventReceiver : EventReceiver<Event.User>
 
 internal class EventProcessorImpl(
     private val eventRepository: EventRepository,
