+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/sync/event/EventProcessorTest.kt
@@ -2,6 +2,8 @@ package com.wire.kalium.logic.sync.event
 
 import com.wire.kalium.logic.data.event.EventRepository
 import com.wire.kalium.logic.framework.TestEvent
+import com.wire.kalium.logic.sync.ConversationEventReceiver
+import com.wire.kalium.logic.sync.UserEventReceiver
 import com.wire.kalium.logic.test_util.TestKaliumDispatcher
 import io.mockative.Mock
 import io.mockative.configure
